minor changes after QA
